Controll : entity 은닉
컨트롤에서 나오는 모든 엔티티, 엔티티화 하는 함수를 모두 Service 객체로 이동, 그에따른 html 참조도 수정

+++ src/main/java/com/example/_2025_bucket/service/TodoService.java
@@ -22,8 +22,8 @@ public class TodoService {
         if(ot.isPresent()){
             Todo todo = ot.get();
             return TodoDto.builder()
-                    .category(todo.getCategory())
-                    .user(todo.getUser())
+                    .categoryDto(todo.getCategory().toDto())
+                    .userDto(todo.getUser().toDto())
                     .reviews(todo.getReviews())
                     .imagePath(todo.getImagePath())
                     .checkComplete(todo.isCheckComplete())
@@ -40,30 +40,31 @@ public class TodoService {
         }
     }
 
-    public List<TodoDto> getAllTodos() {
-        List<Todo> todos = this.todoRepository.findAll(); // Todo 엔티티 리스트 조회
-        List<TodoDto> todoDtos = new ArrayList<>();
-
-        for (Todo todo : todos) {
-            // Todo의 User 필드에서 nickname 추출
-            String nickname = (todo.getUser() != null) ? todo.getUser().getNickname() : "Unknown";
-
-            todoDtos.add(TodoDto.builder()
-                    .category(todo.getCategory())
-                    .user(todo.getUser())
-                    .reviews(todo.getReviews())
-                    .imagePath(todo.getImagePath())
-                    .checkComplete(todo.isCheckComplete())
-                    .modifiedAt(todo.getModifiedAt())
-                    .content(todo.getContent())
-                    .goalDay(todo.getGoalDay())
-                    .uploadAt(todo.getUploadAt())
-                    .id(todo.getId())
-                    .nickname(nickname) // 닉네임 추가
-                    .build());
-        }
-        return todoDtos;
-    }
+    // 어디서 사용하는지 안나와있음, 아마 페이징 추가하면서 안쓰는 것 같음
+//    public List<TodoDto> getAllTodos() {
+//        List<Todo> todos = this.todoRepository.findAll(); // Todo 엔티티 리스트 조회
+//        List<TodoDto> todoDtos = new ArrayList<>();
+//
+//        for (Todo todo : todos) {
+//            // Todo의 User 필드에서 nickname 추출
+//            String nickname = (todo.getUser() != null) ? todo.getUser().getNickname() : "Unknown";
+//
+//            todoDtos.add(TodoDto.builder()
+//                    .categoryDto(todo.getCategory().toDto())
+//                    .userDto(todo.getUser().toDto())
+//                    .reviews(todo.getReviews())
+//                    .imagePath(todo.getImagePath())
+//                    .checkComplete(todo.isCheckComplete())
+//                    .modifiedAt(todo.getModifiedAt())
+//                    .content(todo.getContent())
+//                    .goalDay(todo.getGoalDay())
+//                    .uploadAt(todo.getUploadAt())
+//                    .id(todo.getId())
+//                    .nickname(nickname) // 닉네임 추가
+//                    .build());
+//        }
+//        return todoDtos;
+//    }
 
 
     public void update(TodoDto todoDto) {
@@ -75,32 +76,58 @@ public class TodoService {
     }
 
     public void remove(TodoDto todoDto) {
-        this.todoRepository.delete(todoDto.toEntity());
+        Optional<Todo> ot = this.todoRepository.findById((todoDto.getId()));
+        if(ot.isPresent()){
+            this.todoRepository.delete(todoDto.toEntity());
+        }
+        else {
+            throw new NoSuchElementException("해당 Todo 없음");
+        }
     }
 
-    public List<Todo> getTodosByUserId(long id) {
-        return this.todoRepository.findByUserId(id);
+    public List<TodoDto> getTodosByUserId(long id) {
+        List<Todo> todos =  this.todoRepository.findByUserId(id);
+        List<TodoDto> todoDtos = new ArrayList<>();
+        for(Todo todo : todos){
+            todoDtos.add(TodoDto.builder()
+                            .categoryDto(todo.getCategory().toDto())
+                            .userDto(todo.getUser().toDto())
+                            .reviews(todo.getReviews())
+                            .imagePath(todo.getImagePath())
+                            .checkComplete(todo.isCheckComplete())
+                            .modifiedAt(todo.getModifiedAt())
+                            .content(todo.getContent())
+                            .goalDay(todo.getGoalDay())
+                            .uploadAt(todo.getUploadAt())
+                            .id(todo.getId())
+                            .nickname((todo.getUser() != null) ? todo.getUser().getNickname() : "Unknown")
+                    .build());
+        }
+        return todoDtos;
+
     }
 
-    public List<TodoDto> getTodosByCategory(int categoryId) {
-        List<Todo> todos = todoRepository.findByCategoryId(categoryId); // 카테고리에 해당하는 TODO 조회
-        return todos.stream()
-                .map(todo -> TodoDto.builder()
-                        .category(todo.getCategory())
-                        .user(todo.getUser())
-                        .reviews(todo.getReviews())
-                        .imagePath(todo.getImagePath())
-                        .checkComplete(todo.isCheckComplete())
-                        .modifiedAt(todo.getModifiedAt())
-                        .content(todo.getContent())
-                        .goalDay(todo.getGoalDay())
-                        .uploadAt(todo.getUploadAt())
-                        .id(todo.getId())
-                        .nickname((todo.getUser() != null) ? todo.getUser().getNickname() : "Unknown")
-                        .build())
-                .collect(Collectors.toList());
-    }
+    // 페이징 기능 추가로 미사용
+//    public List<TodoDto> getTodosByCategory(int categoryId) {
+//        List<Todo> todos = todoRepository.findByCategoryId(categoryId); // 카테고리에 해당하는 TODO 조회
+//        return todos.stream()
+//                .map(todo -> TodoDto.builder()
+//                        .categoryDto(todo.getCategory().toDto())
+//                        .userDto(todo.getUser().toDto())
+//                        .reviews(todo.getReviews())
+//                        .imagePath(todo.getImagePath())
+//                        .checkComplete(todo.isCheckComplete())
+//                        .modifiedAt(todo.getModifiedAt())
+//                        .content(todo.getContent())
+//                        .goalDay(todo.getGoalDay())
+//                        .uploadAt(todo.getUploadAt())
+//                        .id(todo.getId())
+//                        .nickname((todo.getUser() != null) ? todo.getUser().getNickname() : "Unknown")
+//                        .build())
+//                .collect(Collectors.toList());
+//    }
 
+    // 카테고리별 페이징 기능
     public Page<TodoDto> getPageTodo(int page, int id){
         List<Sort.Order> sorts = new ArrayList<>();
         sorts.add(Sort.Order.desc("uploadAt"));
@@ -116,8 +143,8 @@ public class TodoService {
 
         List<TodoDto> todoDtos = todos.getContent().stream().map(
                 todo -> TodoDto.builder()
-                        .category(todo.getCategory())
-                        .user(todo.getUser())
+                        .categoryDto(todo.getCategory().toDto())
+                        .userDto(todo.getUser().toDto())
                         .reviews(todo.getReviews())
                         .imagePath(todo.getImagePath())
                         .checkComplete(todo.isCheckComplete())
